feat/jobPost: review middleware

+++ src/services/companies/deleteCompany.service.ts
@@ -1,8 +1,21 @@
 import AppDataSource from '../../data-source'
 import { Company } from '../../entities/companies.entity'
+import AppError from '../../errors/AppError'
 
-const deleteCompanyService = async (companyId: string) => {
+const deleteCompanyService = async (companyId: string, userId: string) => {
     const companyRepository = AppDataSource.getRepository(Company)
+
+    const searchCompany = await companyRepository.findOne({
+        where: {id: companyId},
+        relations: {
+            user: true
+        }
+    })
+
+    if(searchCompany.user.id !== userId) {
+        throw new AppError('This company does not belong', 403)
+    }
+
     await companyRepository.delete({ id: companyId})
 
 }
